+++ packages/validator/index.js
@@ -1,0 +1,5 @@
+const { V, addFeature, removeFeature } = require('src/proxy');
+
+module.exports = {
+    V, addFeature, removeFeature 
+};