+++ packages/validator/src/createStringLengthRangeCheck.js
@@ -1,6 +1,6 @@
 module.exports = function createStringLengthRangeCheck(m, n) {
-    m = m === undefined || m === null ? 0 : m;
-    n = n === undefined || n === null ? Infinity : n;
+    m = (m === undefined || m === null) ? 0 : m;
+    n = (n === undefined || n === null) ? Infinity : n;
     if (typeof n !== 'number') {
             const type = typeof n;
             throw new TypeError(`upper boundery n:<${type}>${n} MUST be of type number`);
@@ -21,13 +21,13 @@ module.exports = function createStringLengthRangeCheck(m, n) {
     if (m < 0) {
             throw new TypeError(`lower boundery m:${m} should be >= 0`);
     }
-    return function isInRange(str) {
+    return function isInRange(str, ...rest) {
             if (typeof str !== 'string'){
-                    return [null, `value type is not of type string: ${typeof str}`];        
+                    return [undefined, `value type is not of type string: ${typeof str}`];        
             }
             if (str.length >= m && str.length <= n) {
-                    return [str, null];
+                    return [[str,...rest], undefined, undefined];
             }
-            return [null, `string of length:${str.length} is not between ${m} and ${n} inclusive`];
+            return [undefined, `string of length:${str.length} is not between ${m} and ${n} inclusive`];
     }
 }