+++ packages/validator/CHANGELOG.md
@@ -1,4 +1,10 @@
+##  0.0.2 (unreleased)
+- 'url' feature added and tested
+- xref uses @mangos/jxpath/internals
+- proxy is re-written to have different handlers depending on what state the validator is in during construction
+- all regression tests now working for the re-write
+
 ## 0.0.1
 
-Released August 7th, 2019
+Released dec 28, 2019
  - initial release